requirements.txt
Package                  Version
------------------------ ----------
aiohappyeyeballs         2.6.1
aiohttp                  3.13.2
aiosignal                1.4.0
annotated-types          0.7.0
anyio                    4.12.0
attrs                    25.4.0
certifi                  2025.11.12
charset-normalizer       3.4.4
dataclasses-json         0.6.7
distro                   1.9.0
edge-tts                 7.2.3
frozenlist               1.8.0
greenlet                 3.3.0
groq                     0.37.1
h11                      0.16.0
httpcore                 1.0.9
httpx                    0.28.1
httpx-sse                0.4.3
idna                     3.11
jiter                    0.12.0
jsonpatch                1.33
jsonpointer              3.0.0
langchain                1.1.2
langchain-classic        1.0.0
langchain-community      0.4.1
langchain-core           1.1.1
langchain-groq           1.1.0
langchain-openai         1.1.0
langchain-text-splitters 1.0.0
langchain-tools          0.1.34
langgraph                1.0.4
langgraph-checkpoint     3.0.1
langgraph-prebuilt       1.0.5
langgraph-sdk            0.2.14
langsmith                0.4.56
marshmallow              3.26.1
multidict                6.7.0
mypy_extensions          1.1.0
numpy                    2.3.5
openai                   2.9.0
orjson                   3.11.5
ormsgpack                1.12.0
packaging                25.0
pip                      25.1.1
propcache                0.4.1
pydantic                 2.12.5
pydantic_core            2.41.5
pydantic-settings        2.12.0
pygame                   2.6.1
python-decouple          3.8
python-dotenv            1.2.1
PyYAML                   6.0.3
regex                    2025.11.3
requests                 2.32.5
requests-toolbelt        1.0.0
sniffio                  1.3.1
SpeechRecognition        3.14.4
SQLAlchemy               2.0.44
tabulate                 0.9.0
tenacity                 9.1.2
tiktoken                 0.12.0
tqdm                     4.67.1
typing_extensions        4.15.0
typing-inspect           0.9.0
typing-inspection        0.4.2
urllib3                  2.6.0
uuid_utils               0.12.0
xxhash                   3.6.0
yarl                     1.22.0
zstandard                0.25.0
